% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{calc_dens_par}
\alias{calc_dens_par}
\title{Calculate the value of the density function (parametric form)}
\usage{
calc_dens_par(time, dist, pars)
}
\arguments{
\item{time}{is the time at which the distribution function should be calculated.}

\item{dist}{is the statistical distribution (named per \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}).}

\item{pars}{is a vector of the parameters for that distribution.
\itemize{
\item Exponential distribution (\code{exp}) requires the rate parameter.
\item Weibull distribution (both \code{weibullPH} and \code{weibull} formulations) requires the shape and scale parameters.
\item Log-logistic distribution (\code{llogis}) requires the shape and scale parameters.
\item Log-normal distribution (\code{lnorm}) requires the meanlog and sdlog parameters.
\item Gamma and Gompertz distributions (\code{gamma} and \code{gompertz}) require the shape and rate parameters.
\item Generalized Gamma requires the mu, sigma and Q parameters if using the standard parameterization (\code{gengamma}) or shape, scale and k parameters if using the original parameterization (\code{gengamma.orig}).
}}
}
\value{
The value of a density function, a numeric value.
}
\description{
Calculate the value of the density function, given the statistical distribution and its parameters.
}
\examples{
calc_dens_par(10, "exp", 0.01)
calc_dens_par(5, "lnorm", c(3, 1))
}
\seealso{
\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}
}
