% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{calc_ibs}
\alias{calc_ibs}
\title{Calculate Integrated Brier Scores for Overall Survival}
\usage{
calc_ibs(ptdata, dpam)
}
\arguments{
\item{ptdata}{Patient-level dataset}

\item{dpam}{List of survival regressions for model endpoints. Should include all six endpoints.}
}
\value{
List containing
\itemize{
\item osibs: Vector of integrated Brier scores for PSM, STM-CF and STM-CR decision models
\item timerange: Vector of the time range over which the integrated Brier score is calculated
}
}
\description{
Calculate integrated Brier score for overall survival for each decision model. This function is essentially a wrapper to \link[SurvMetrics:IBS]{SurvMetrics::IBS} for use with this package.
}
\examples{
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_par(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_par(fits$ppd, "aic")$fit,
  ttp = find_bestfit_par(fits$ttp, "aic")$fit,
  pfs = find_bestfit_par(fits$pfs, "aic")$fit,
  os = find_bestfit_par(fits$os, "aic")$fit,
  pps_cf = find_bestfit_par(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_par(fits$pps_cr, "aic")$fit
  )
calc_ibs(bosonc, params)
}
