% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltablesurv.R
\name{calc_ltsurv}
\alias{calc_ltsurv}
\title{Calculate survival from a lifetable}
\usage{
calc_ltsurv(time, lifetable)
}
\arguments{
\item{time}{The time(s) to which survival is to be estimated (from time zero).}

\item{lifetable}{The lifetable must be a dataframe with columns named time and lx. The first entry of the time column must be zero. Data should be sorted in ascending order by time, and all times must be unique.}
}
\value{
Numeric survival probability
}
\description{
Calculate survival from time zero to a given time, according to a provided lifetable
}
\examples{
ltable <- tibble::tibble(time=0:10, lx=1-time*0.1)
calc_ltsurv(c(2, 2.5, 9.3), ltable)
}
