% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmeans.R
\name{calc_rmd_first}
\alias{calc_rmd_first}
\title{Calculate restricted mean duration in respect of the first part of a two-piece model
Internal function to calculate the restricted mean duration up to the cut-off time, the first part of a two-piece model. Assumes the time horizon is beyond the cutoff time.}
\usage{
calc_rmd_first(ds, cuttime)
}
\arguments{
\item{ds}{patient-level dataset}

\item{cuttime}{time cut-off}
}
\value{
list containing:
\itemize{
\item pfsurv: the PF survival function at the cut-off time
\item pfarea: area under the PF survival function up to the cut-off time
\item ossurv: the OS survival function at the cut-off time
\item osarea: area under the OS survival function up to the cut-off time
}
}
\description{
Calculate restricted mean duration in respect of the first part of a two-piece model
Internal function to calculate the restricted mean duration up to the cut-off time, the first part of a two-piece model. Assumes the time horizon is beyond the cutoff time.
}
\examples{
cuttime <- 20
bosonc <- create_dummydata("flexbosms")
ds <- create_extrafields(bosonc, cuttime=cuttime)
dpam <- fit_ends_mods_par(bosonc, cuttime=cuttime)
calc_rmd_first(ds=ds, cuttime=cuttime)
}
