% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{calc_rmd_par}
\alias{calc_rmd_par}
\title{Calculate restricted mean durations (parametric form)}
\usage{
calc_rmd_par(Tw, dist, pars)
}
\arguments{
\item{Tw}{is the time period over which the restricted mean is calculated}

\item{dist}{is the name (from flexsurv) of a particular statistical distribution, e.g. "llogis" is the log-logistic distribution.}

\item{pars}{is a vector whose components are the parameters for the statistical distribution (see note).}
}
\value{
the restricted mean duration, a numeric value.
}
\description{
Calculates the restricted mean duration, given a statistical distribution (named per flexsurv) and its parameters.
}
\examples{
calc_rmd_par("exp", 0.2, 20)
calc_rmd_par("lnorm", c(3, 1), 10)
}
\seealso{
\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} and \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}
}
