% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{check_pfsos_consistent}
\alias{check_pfsos_consistent}
\title{Check the consistency of the chosen PFS and OS distributions}
\usage{
check_pfsos_consistent(dpam, Ty = NA)
}
\arguments{
\item{dpam}{List of survival regressions for model endpoints. This must include progression-free survival (PFS) and overall survival (OS).}

\item{Ty}{Time duration over which to calculate. Assumes input is in years, and patient-level data is recorded in weeks. NA implies there is no maximum.}
}
\value{
List containing:
\itemize{
\item check: TRUE if check passes, FALSE if check fails
\item mess: provides and error message if the check fails
}
}
\description{
Check the consistency of the chosen PFS and OS distributions
}
\examples{
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_par(bosonc)
# Pick out best distributions for PFS and OS according to min AIC
params_aic <- list(
  pfs = find_bestfit_par(fits$pfs, "aic")$fit,
  os = find_bestfit_par(fits$os, "aic")$fit
)
params_exp <- list(
  pfs = fits$pfs[[1]]$result,
  os = fits$os[[1]]$result
)
check_pfsos_consistent(params_aic)
check_pfsos_consistent(params_exp, Ty=10)
}
