% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{check_posdef}
\alias{check_posdef}
\title{Checks whether the Hessian for a survival regression is positive-definite}
\usage{
check_posdef(fitlist)
}
\arguments{
\item{fitlist}{is a list returned after running \code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} describing a fitted survival model}
}
\value{
logical: TRUE if Hessian matrix is positive definite, FALSE if not.
}
\description{
Checks whether the Hessian matrix returned in a list after fitting a survival regression with flexsurvreg is positive-definite.
}
\examples{
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_par(bosonc)
check_posdef(fits$pfs[[2]]$result)
}
