% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltablesurv.R
\name{constrain_survprob}
\alias{constrain_survprob}
\title{Constrain survival probabilities according to hazards in a lifetable
Recalculated constrained survival probabilities (by week) as the lower of the original unadjusted survival probability and the survival implied by the given lifetable (assumed indexed as years).}
\usage{
constrain_survprob(survprob, lifetable, timevec = 0:(length(survprob) - 1))
}
\arguments{
\item{survprob}{}

\item{lifetable}{}

\item{timevec}{}
}
\value{
Vector of constrained survival probabilities
}
\description{
Constrain survival probabilities according to hazards in a lifetable
Recalculated constrained survival probabilities (by week) as the lower of the original unadjusted survival probability and the survival implied by the given lifetable (assumed indexed as years).
}
\examples{
ltable <- tibble::tibble(lttime=0:20, lx=c(1,0.08,0.05,0.03,0.01,rep(0,16)))
survprob <- c(1,0.5,0.4,0.2,0)
constrain_survprob(survprob, lifetable=ltable)
timevec <- 100*(0:4)
constrain_survprob(survprob, lifetable=ltable, timevec=timevec)
}
