% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{convert_wks2yrs}
\alias{convert_wks2yrs}
\title{Convert weeks to years}
\usage{
convert_wks2yrs(weeks)
}
\arguments{
\item{weeks}{Number of weeks}
}
\value{
Number of years = weeks x 7 / 365.25
}
\description{
Convert weeks to years. Function allows hard-coding to be minimized elsewhere in the package.
}
\seealso{
\link{convert_yrs2wks}
}
