% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{convert_yrs2wks}
\alias{convert_yrs2wks}
\title{Convert years to weeks}
\usage{
convert_yrs2wks(years)
}
\arguments{
\item{years}{Number of years}
}
\value{
Number of weeks = years / 7 * 365.25
}
\description{
Convert weeks to years. Function allows hard-coding to be minimized elsewhere in the package.
}
\seealso{
\link{convert_wks2yrs}
}
