% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{create_dummydata_flexbosms}
\alias{create_dummydata_flexbosms}
\title{Create flexbosms dataset for illustration}
\usage{
create_dummydata_flexbosms()
}
\value{
Tibble dataset, for use with \link{psm3mkv} functions
}
\description{
Create 'flexbosms' dummy dataset to illustrate \link{psm3mkv}
}
\seealso{
\code{\link[=create_dummydata]{create_dummydata()}}
}
