% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrmd.R
\name{discretized_rmd}
\alias{discretized_rmd}
\title{Restricted mean duration in progression-free for state transition models, discretized approximation}
\usage{
discretized_rmd(dpam, Ty = 10, discrate = 0, state, model, timestep = 1)
}
\arguments{
\item{dpam}{List of survival regressions for model endpoints. These must include time to progression (TTP) and pre-progression death (PPD).}

\item{Ty}{Time duration over which to calculate. Assumes input is in years, and patient-level data is recorded in weeks.}

\item{discrate}{Discount rate (\%) per year}

\item{state}{May be "PF", "PD" or "OS" being the progression-free or progressive disease states, or overall survival}

\item{model}{May be "PSM", "STM-CF" or "STM-CR" model structures}

\item{starting}{Vector of membership probabilities at time zero.}

\item{lifetable}{Optional. The lifetable must be a dataframe with columns named time and lx. The first entry of the time column must be zero. Data should be sorted in ascending order by time, and all times must be unique.}
}
\value{
Numeric value in same time unit as patient-level data (weeks).
}
\description{
Calculates a discretized approximation for the mean duration in the progression-free state for both the state transition clock forward and clock reset models. Requires a carefully formatted list of fitted survival regressions for the necessary endpoints, and the time duration to calculate over.
This method is only valid for one-piece models, without lifetable adjustment.
}
\examples{
# Create dataset and fit survival models (splines)
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_spl(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_spl(fits$ppd, "aic")$fit,
  ttp = find_bestfit_spl(fits$ttp, "aic")$fit,
  pfs = find_bestfit_spl(fits$pfs, "aic")$fit,
  os = find_bestfit_spl(fits$os, "aic")$fit,
  pps_cf = find_bestfit_spl(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_spl(fits$pps_cr, "aic")$fit
)
discretized_rmd(dpam=params, state="pd", model="stm_cr")
rmd_pd_stm_cr(dpam=params)
}
\seealso{
Full integral calculation in \code{\link[=prmd_pf_stm]{prmd_pf_stm()}}
}
