% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{find_bestfit_par}
\alias{find_bestfit_par}
\title{Find the "best" survival regression from a list}
\usage{
find_bestfit_par(reglist, crit)
}
\arguments{
\item{reglist}{List of fitted survival regressions to an endpoint and dataset.}

\item{crit}{Criterion to be used in selection of best fit, either "aic" (Akaike Information Criterion) or "bic" (Bayesian Information Criterion).}
}
\value{
List of the single survival regression with the best fit.
}
\description{
When there are multiple survival regressions fitted to the same endpoint and dataset, it is necessary to identify the preferred model. This function reviews the fitted regressions and selects that with the minimum Akaike or Bayesian Information Criterion (AIC, BIC), depending on user choice.
}
\examples{
bosonc <- create_dummydata("flexbosms")
# Pick some distributions to fit to all endpoints
dists <- c("exp", "llogis", "lnorm")
# Fit all distributions to all endpoints
fits <- fit_ends_mods_par(bosonc,
    cuttime = 0,
    ppd.dist = dists,
    ttp.dist = dists,
    pfs.dist = dists,
    os.dist = dists,
    pps_cr.dist = dists,
    pps_cf.dist = dists)
find_bestfit_par(fits$ttp, "aic")
find_bestfit_par(fits$os, "bic")
}
