% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmeans.R
\name{fit_ends_mods_given}
\alias{fit_ends_mods_given}
\title{Fit survival models to each endpoint, given type and spec
Internal function to fit survival models to each endpoint, given type and spec in format of \code{\link[=convert_fit2spec]{convert_fit2spec()}}}
\usage{
fit_ends_mods_given(simdat, dpam, cuttime)
}
\arguments{
\item{simdat}{is the (sample of the) patient-level dataset}

\item{dpam}{is the currently fitted set of survival models to each endpoint}

\item{cuttime}{is the cut-off time for two-piece modeling}
}
\value{
A list by endpoint, then distribution, each containing two components:
\itemize{
\item result: A list of class \emph{flexsurvreg} containing information about the fitted model.
\item error: Any error message returned on fitting the regression (NULL indicates no error).
}
}
\description{
Fit survival models to each endpoint, given type and spec
Internal function to fit survival models to each endpoint, given type and spec in format of \code{\link[=convert_fit2spec]{convert_fit2spec()}}
}
\seealso{
\code{\link[=convert_fit2spec]{convert_fit2spec()}}, \code{\link[=fit_ends_mods_par]{fit_ends_mods_par()}}, \code{\link[=fit_ends_mods_spl]{fit_ends_mods_spl()}}
}
