% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_mods_par}
\alias{fit_mods_par}
\title{Fit survival regressions with multiple parametric distributions}
\usage{
fit_mods_par(durn1, durn2 = NA, evflag, dists, expvar = NA)
}
\arguments{
\item{durn1}{First time point, corresponds with time in \code{\link[survival:Surv]{survival::Surv()}}. For right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval.}

\item{durn2}{Second time point, corresponds with time2 in \code{\link[survival:Surv]{survival::Surv()}}. The ending time of the interval for interval censored or counting process data only.}

\item{evflag}{Event flag, corresponds with event in \link[survival:Surv]{survival::Surv}. The status indicator is normally 0=alive, 1=dead.}

\item{dists}{is a vector of all the distributions (named according to \code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}}) for which statistical fits are requested.}

\item{expvar}{Explanatory variable for modeling of PPS}
}
\value{
A list by distribution, each containing two components:
\itemize{
\item result: A list of class \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg} containing information about the fitted model.
\item error: Any error message returned on fitting the regression (NULL indicates no error).
}
}
\description{
Fits survival regressions with \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg} to multiple statistical distributions in one function call.
}
