% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-spl.R
\name{fit_mods_spl}
\alias{fit_mods_spl}
\title{Fit survival regressions with multiple spline regressions}
\usage{
fit_mods_spl(
  durn1,
  durn2 = NA,
  evflag,
  knot_set = 1:3,
  scale_set = c("hazard", "odds", "normal"),
  expvar = NA
)
}
\arguments{
\item{durn1}{First time point, corresponds with time in \code{\link[survival:Surv]{survival::Surv()}}. For right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval.}

\item{durn2}{Second time point, corresponds with time2 in \code{\link[survival:Surv]{survival::Surv()}}. The ending time of the interval for interval censored or counting process data only.}

\item{evflag}{Event flag, corresponds with event in \link[survival:Surv]{survival::Surv}. The status indicator is normally 0=alive, 1=dead.}

\item{knot_set}{is a vector of the numbers of knots to consider, following \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}}).}

\item{scale_set}{is a vector of the spline scales to consider, following \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}}).}

\item{expvar}{Explanatory variable for modeling of PPS}
}
\value{
A list by distribution, each containing two components:
\itemize{
\item result: A list of class \link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} containing information about the fitted model.
\item error: Any error message returned on fitting the regression (NULL indicates no error).
}
}
\description{
Fits survival regressions with \link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} to multiple Royston-Parmar splines specifications in one function call.
}
\seealso{
\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} and \code{\link[survival:Surv]{survival::Surv()}}
}
