% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{give_noparams}
\alias{give_noparams}
\title{Number of parameters used by parametric statistical distributions}
\usage{
give_noparams(type, spec)
}
\arguments{
\item{type}{is either "par" for regular parametric form (exponential, weibull etc) or "spl" for Royston-Parmar splines.}

\item{spec}{is a list comprising:
If type=="par": \code{dist} is the statistical distribution (named per \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}) and \code{pars} is a vector of the parameters for that distribution.
\itemize{
\item Exponential distribution (\code{exp}) requires the rate parameter.
\item Weibull distribution (both \code{weibullPH} and \code{weibull} formulations) requires the shape and scale parameters.
\item Log-logistic distribution (\code{llogis}) requires the shape and scale parameters.
\item Log-normal distribution (\code{lnorm}) requires the meanlog and sdlog parameters.
\item Gamma and Gompertz distributions (\code{gamma} and \code{gompertz}) require the shape and rate parameters.
\item Generalized Gamma requires the mu, sigma and Q parameters if using the standard parameterization (\code{gengamma}) or shape, scale and k parameters if using the original parameterization (\code{gengamma.orig}).
If type=="spl":
\item \code{gamma} - Vector of parameters describing the baseline spline function, as described in \link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline}. This may be supplied as a vector with number of elements equal to the length of knots, in which case the parameters are common to all times. Alternatively a matrix may be supplied, with rows corresponding to different times, and columns corresponding to knots.
\item \code{knots} - Vector of locations of knots on the axis of log time, supplied in increasing order. Unlike in \link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline}, these include the two boundary knots.
\item \code{scale} - Either "hazard", "odds", or "normal", as described in \link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline}. With the default of no knots in addition to the boundaries, this model reduces to the Weibull, log-logistic and log-normal respectively. The scale must be common to all times.
}}
}
\value{
a numeric value
}
\description{
Returns the number of parameters used by the specified statistical model
}
\examples{
give_noparams(type="par", spec=list(dist="weibullPH"))
give_noparams(type="spl", spec=list(gamma=c(1.1,2.1,3.1)))
}
\seealso{
\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} and \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}
}
