% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{give_noparams_par}
\alias{give_noparams_par}
\title{Number of parameters used by parametric statistical distributions}
\usage{
give_noparams_par(dist)
}
\arguments{
\item{dist}{is the statistical distribution (named per \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg})}
}
\value{
a numeric value
}
\description{
Returns the number of parameters used by one or many statistical distributions, named as per flexsurv.
}
\examples{
give_noparams_par("llogis")
}
\seealso{
\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} and \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}
}
