% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdpps.R
\name{graph_psm_hazards}
\alias{graph_psm_hazards}
\title{Graph the PSM hazard functions}
\usage{
graph_psm_hazards(timevar, endpoint, ptdata, dpam, psmtype)
}
\arguments{
\item{timevar}{Vector of times at which to calculate the hazards}

\item{endpoint}{Endpoint for which hazard is required (TTP, PPD, PFS, OS or PPS)}

\item{ptdata}{Patient-level dataset}

\item{dpam}{List of fitted survival models for each endpoint}

\item{psmtype}{Type of PSM - simple or complex}
}
\value{
List of data and graph
Data is a dataset of the time-range, method (adjusted or unadjusted hazard), and hazard value
Graph is a line graphic of the hazards plotted against the time-range
}
\description{
EXPERIMENTAL. Graph the PSM hazard functions
}
\examples{
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_par(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_par(fits$ppd, "aic")$fit,
  ttp = find_bestfit_par(fits$ttp, "aic")$fit,
  pfs = find_bestfit_par(fits$pfs, "aic")$fit,
  os = find_bestfit_par(fits$os, "aic")$fit,
  pps_cf = find_bestfit_par(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_par(fits$pps_cr, "aic")$fit
)
# Create graphics
# psmh_simple <- graph_psm_hazards(
#   timerange=(0:10)*6,
#   endpoint="OS",
#   ptdata=bosonc,
#   dpam=params,
#   psmtype="simple")
# psmh_simple$graph
}
