% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probgraphs.R
\name{graph_survorhaz}
\alias{graph_survorhaz}
\title{Graph the observed and fitted state membership probabilities}
\usage{
graph_survorhaz(ptdata, dpam, metric = "survival", cuttime = 0, tpoints = 100)
}
\arguments{
\item{ptdata}{is the patient-level dataset of TTP, PFS and OS.}

\item{dpam}{is the list of survival regressions for model endpoints. All endpoints are required.}

\item{cuttime}{is the cut-off time for a two-piece model (default 0, indicating a one-piece model)}

\item{tpoints}{indicates how many timepoints should be included in the graphics (default 100)}
}
\value{
Four datasets and graphics as a list
}
\description{
Graph the observed and fitted state membership probabilities for PF, PD, OS and PPS.
}
\examples{
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_spl(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_spl(fits$ppd, "aic")$fit,
  ttp = find_bestfit_spl(fits$ttp, "aic")$fit,
  pfs = find_bestfit_spl(fits$pfs, "aic")$fit,
  os = find_bestfit_spl(fits$os, "aic")$fit,
  pps_cf = find_bestfit_spl(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_spl(fits$pps_cr, "aic")$fit
)
# Create graphics
ptdgraphs <- graph_survs(bosonc, params)
ptdgraphs$data$pps[1:10,]
ptdgraphs$graph$pf
}
