% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probgraphs.R
\name{graph_survs_2p}
\alias{graph_survs_2p}
\title{Graph the observed and fitted state membership probabilities}
\usage{
graph_survs_2p(ptdata, dpam, cuttime = 0, tpoints = 100)
}
\arguments{
\item{ptdata}{is the patient-level dataset of TTP, PFS and OS.}

\item{dpam}{is the list of survival regressions for model endpoints. All endpoints are required.}

\item{cuttime}{is the cut-off time for a two-piece model (default 0, indicating a one-piece model)}

\item{tpoints}{indicates how many timepoints should be included in the graphics (default 100)}
}
\value{
Four datasets and graphics as a list
}
\description{
Graph the observed and fitted state membership probabilities for PF, PD, OS and PPS.
}
