% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmeans.R
\name{prmd_os_psm}
\alias{prmd_os_psm}
\title{Safely calculate restricted mean duration for overall survival in the partitioned survival model}
\usage{
prmd_os_psm(...)
}
\arguments{
\item{...}{Pass-through to \link{rmd_os_psm}}
}
\value{
Numeric value in same time unit as patient-level data (weeks).
}
\description{
Calculates the mean duration alive (i.e. in either progression free or progressed disease states) for the partitioned survival model. Requires a carefully formatted list of fitted survival regressions for necessary endpoints, and the time duration to calculate over. Wrapper with 'possibly' of \link{rmd_os_psm}. This function is called by \link{calc_allrmds}.
}
