% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmeans.R
\name{prmd_pf_psm}
\alias{prmd_pf_psm}
\title{Safely calculate restricted mean duration in progression free state for the partitioned survival model}
\usage{
prmd_pf_psm(...)
}
\arguments{
\item{...}{Pass-through to \link{rmd_pf_psm}}
}
\value{
Numeric value in same time unit as patient-level data (weeks).
}
\description{
Calculates the mean duration in the progression free state for the partitioned survival model. Requires a carefully formatted list of fitted survival regressions for necessary endpoints, and the time duration to calculate over. Wrapper with 'possibly' of \link{rmd_pf_psm}. This function is called by \link{calc_allrmds}.
}
