% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmeans.R
\name{prmd_pf_stm}
\alias{prmd_pf_stm}
\title{Safely calculate restricted mean duration in progression-free for state transition models}
\usage{
prmd_pf_stm(...)
}
\arguments{
\item{...}{Pass-through to \link{rmd_pf_stm}}
}
\value{
Numeric value in same time unit as patient-level data (weeks).
}
\description{
Calculates the mean duration in the progression-free state for both the state transition clock forward and clock reset models. Requires a carefully formatted list of fitted survival regressions for the necessary endpoints, and the time duration to calculate over. Wrapper with 'possibly' of \link{rmd_pf_stm}. This function is called by \link{calc_allrmds}.
}
