% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltablesurv.R
\name{vonelookup}
\alias{vonelookup}
\title{VoneLOOKUP function}
\usage{
vonelookup(oneindexval, indexvec, valvec, method = "geom")
}
\arguments{
\item{oneindexval}{The single index value to be looked-up}

\item{indexvec}{The vector of indices to look-up in}

\item{valvec}{The vector of values corresponding to the vector of indices}

\item{method}{Method may be \code{floor}, \code{ceiling}, \code{arith} or \code{geom} (default).}
}
\value{
Numeric value or vector, depending on the lookup/interpolation method chosen:
\itemize{
\item \code{floor}: Floor value, where interpolation is required between measured values
\item \code{ceiling}: Ceiling value, where interpolation is required between measured values
\item \code{arith}: Arithmetic mean, where interpolation is required between measured values
\item \code{geom}: Geometric mean, where interpolation is required between measured values
}
}
\description{
Function to lookup a single (one) value according to an index. Aims to behave similarly to VLOOKUP in Microsoft Excel.
}
\seealso{
\link{vlookup}
}
