% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhoods.R
\name{calc_likes}
\alias{calc_likes}
\title{Calculate likelihoods for three three-state model structures}
\usage{
calc_likes(ptdata, dpam, cuttime = 0)
}
\arguments{
\item{ptdata}{Dataset of patient level data. Must be a tibble with columns named:
\itemize{
\item ptid: patient identifier
\item pfs.durn: duration of PFS from baseline
\item pfs.flag: event flag for PFS (=1 if progression or death occurred, 0 for censoring)
\item os.durn: duration of OS from baseline
\item os.flag: event flag for OS (=1 if death occurred, 0 for censoring)
\item ttp.durn: duration of TTP from baseline (usually should be equal to pfs.durn)
\item ttp.flag: event flag for TTP (=1 if progression occurred, 0 for censoring).
}

Survival data for all other endpoints (time to progression, pre-progression death, post-progression survival) are derived from PFS and OS.}

\item{dpam}{List of survival regressions for each endpoint:
\itemize{
\item pre-progression death (PPD)
\item time to progression (TTP)
\item progression-free survival (PFS)
\item overall survival (OS)
\item post-progression survival clock forward (PPS-CF) and
\item post-progression survival clock reset (PPS-CR).
}}

\item{cuttime}{Time cutoff - this is nonzero for two-piece models.}
}
\value{
Two outputs are returned:

\code{results} is a tibble of values and data relating to the likelihood for this model:
\itemize{
\item \code{npts}: Number of patients analysed for each endpoint.
\item \code{likedata}: Patient-level dataset with additional likelihood-related calculations.
\item \code{coefsdists}: Summary table of distributions and parameters used for each endpoint.
\item \code{slikes}: Total log-likelihood for each possible outcome
\item \code{ll}: Total log-likelihood
\item \code{params}: Number of parameters used in this model
\item \code{AIC}: Akaike Information Criterion value for this model
\item \code{BIC}: Bayesian Information Criterion value for this model
}

\code{llcomp} is a tibble providing a breakdown of the likelihood calculations by outcome. Outcomes are as follows:
\itemize{
\item (1) refers to patients who remain alive and progression-free during the follow-up;
\item (2) refers to patients who die without prior progression during the follow-up;
\item (3) refers to patients who progress and then remain alive for the remaining follow-up, and
\item (4) refers to patients who progress and die within the follow-up.
}

The number of patients for each outcome are given for each model structure. You may confirm that these are identical across model structures.
The contribution of each patient group to the calculation of log-likelihood for each model is given in fields beginning \code{ll_}. This is helpful for understanding differences in likelihoods between model structures, according to patient outcomes.
}
\description{
Calculate likelihood values and other summary output for the following three state models structures: partitioned survival, clock forward state transition, and clock reset state transition. The function requires appropriately formatted patient-level data, a set of fitted survival regressions, and the time cut-off (if two-piece modeling is used).
}
\examples{
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_spl(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_spl(fits$ppd, "aic")$fit,
  ttp = find_bestfit_spl(fits$ttp, "aic")$fit,
  pfs = find_bestfit_spl(fits$pfs, "aic")$fit,
  os = find_bestfit_spl(fits$os, "aic")$fit,
  pps_cf = find_bestfit_spl(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_spl(fits$pps_cr, "aic")$fit
  )
calc_likes(bosonc, dpam=params)
}
\seealso{
This function calls \code{\link[=calc_likes_psm_simple]{calc_likes_psm_simple()}}, \code{\link[=calc_likes_psm_complex]{calc_likes_psm_complex()}}, \code{\link[=calc_likes_stm_cf]{calc_likes_stm_cf()}}, and \code{\link[=calc_likes_stm_cr]{calc_likes_stm_cr()}}.
}
