% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{calc_rmd_par}
\alias{calc_rmd_par}
\title{Calculate restricted mean durations (parametric form)}
\usage{
calc_rmd_par(Tw, dist, pars)
}
\arguments{
\item{Tw}{is the time period over which the restricted mean is calculated}

\item{dist}{is the statistical distribution (named per \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}).}

\item{pars}{is a vector of the parameters for that distribution.
\itemize{
\item Exponential distribution (\code{exp}) requires the rate parameter.
\item Weibull distribution (both \code{weibullPH} and \code{weibull} formulations) requires the shape and scale parameters.
\item Log-logistic distribution (\code{llogis}) requires the shape and scale parameters.
\item Log-normal distribution (\code{lnorm}) requires the meanlog and sdlog parameters.
\item Gamma and Gompertz distributions (\code{gamma} and \code{gompertz}) require the shape and rate parameters.
\item Generalized Gamma requires the mu, sigma and Q parameters if using the standard parameterization (\code{gengamma}) or shape, scale and k parameters if using the original parameterization (\code{gengamma.orig}).
}}
}
\value{
the restricted mean duration, a numeric value.
}
\description{
Calculates the restricted mean duration, given a parametric statistical distribution.
}
\examples{
calc_rmd_par(20, "exp", 0.2)
calc_rmd_par(10, "lnorm", c(3, 1))
}
\seealso{
\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline} and \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}
}
