% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{create_extrafields}
\alias{create_extrafields}
\title{Create the additional time-to-event endpoints, adjusting for cutpoint}
\usage{
create_extrafields(ds, cuttime = 0)
}
\arguments{
\item{ds}{Patient-level dataset}

\item{cuttime}{Time cutpoint}
}
\value{
Dataset of complete patient-level dataset, adjusted for cutpoint
}
\description{
Create the additional time-to-event endpoints, adjusting for cutpoint
}
\examples{
bosonc <- create_dummydata("flexbosms")
create_extrafields(bosonc, cuttime=10)
}
