% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrmd.R
\name{drmd_psm}
\alias{drmd_psm}
\title{Discretized Restricted Mean Duration calculation for Partitioned Survival Model
Calculate restricted mean duration (RMD) in PF, PD and OS states under a Partitioned Survival Model structure.}
\usage{
drmd_psm(
  ptdata,
  dpam,
  psmtype = "simple",
  Ty = 10,
  discrate = 0,
  lifetable = NA,
  timestep = 1
)
}
\arguments{
\item{ptdata}{Dataset of patient level data. Must be a tibble with columns named:
\itemize{
\item ptid: patient identifier
\item pfs.durn: duration of PFS from baseline
\item pfs.flag: event flag for PFS (=1 if progression or death occurred, 0 for censoring)
\item os.durn: duration of OS from baseline
\item os.flag: event flag for OS (=1 if death occurred, 0 for censoring)
\item ttp.durn: duration of TTP from baseline (usually should be equal to pfs.durn)
\item ttp.flag: event flag for TTP (=1 if progression occurred, 0 for censoring).
}

Survival data for all other endpoints (time to progression, pre-progression death, post-progression survival) are derived from PFS and OS.}

\item{dpam}{List of survival regressions for model endpoints. These must include time to progression (TTP) and pre-progression death (PPD).}

\item{psmtype}{Either "simple" or "complex" PSM formulation}

\item{Ty}{Time duration over which to calculate (defaults to 10 years). Assumes input is in years, and patient-level data is recorded in weeks.}

\item{discrate}{Discount rate (\%) per year (defaults to zero).}

\item{lifetable}{Optional. The lifetable must be a dataframe with columns named time and lx. The first entry of the time column must be zero. Data should be sorted in ascending order by time, and all times must be unique.}

\item{timestep}{Optional, defaults to one (week).}
}
\value{
List containing:
\itemize{
\item pf: RMD in PF state
\item pd: RMD in PD state
\item os: RMD in either alive state
}
}
\description{
Discretized Restricted Mean Duration calculation for Partitioned Survival Model
Calculate restricted mean duration (RMD) in PF, PD and OS states under a Partitioned Survival Model structure.
}
\examples{
# Create dataset and fit survival models (splines)
bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_spl(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_spl(fits$ppd, "aic")$fit,
  ttp = find_bestfit_spl(fits$ttp, "aic")$fit,
  pfs = find_bestfit_spl(fits$pfs, "aic")$fit,
  os = find_bestfit_spl(fits$os, "aic")$fit,
  pps_cf = find_bestfit_spl(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_spl(fits$pps_cr, "aic")$fit
)
drmd_psm(ptdata=bosonc, dpam=params)
# Add a lifetable constraint
ltable <- tibble::tibble(lttime=0:20, lx=1-lttime*0.05)
drmd_psm(ptdata=bosonc, dpam=params, lifetable=ltable)
}
\seealso{
\code{\link[=drmd_stm_cf]{drmd_stm_cf()}} \code{\link[=drmd_stm_cr]{drmd_stm_cr()}}
}
