% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdpps.R
\name{pickout_psmhaz}
\alias{pickout_psmhaz}
\title{Obtain adjusted and unadjusted PSM hazards}
\usage{
pickout_psmhaz(timepoint, endpoint, ptdata, dpam, psmtype)
}
\arguments{
\item{endpoint}{Endpoint for which hazard is required (TTP, PPD, PFS, OS or PPS)}

\item{ptdata}{Patient-level dataset}

\item{dpam}{List of fitted survival models for each endpoint}

\item{psmtype}{Type of PSM - simple or complex}
}
\value{
List of data and graph
Data is a dataset of the time-range, method (adjusted or unadjusted hazard), and hazard value
Graph is a line graphic of the hazards plotted against the time-range

adj is the hazard adjusted for constraints, unadj is the unadjusted hazard
}
\description{
EXPERIMENTAL. Obtain adjusted and unadjusted PSM hazards for given endpoint and time
}
