% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdpps.R
\name{pickout_psmhaz}
\alias{pickout_psmhaz}
\title{Obtain adjusted and unadjusted PSM hazards}
\usage{
pickout_psmhaz(timevar, endpoint = NA, dpam, psmtype)
}
\arguments{
\item{timevar}{Vector of times at which to calculate the hazards}

\item{endpoint}{Endpoint for which hazard is required (TTP, PPD, PFS, OS or PPS)}

\item{dpam}{List of survival regressions for each endpoint:
\itemize{
\item pre-progression death (PPD)
\item time to progression (TTP)
\item progression-free survival (PFS)
\item overall survival (OS)
\item post-progression survival clock forward (PPS-CF) and
\item post-progression survival clock reset (PPS-CR).
}}

\item{psmtype}{Type of PSM - simple or complex}
}
\value{
\code{adj} is the hazard adjusted for constraints, \code{unadj} is the unadjusted hazard
}
\description{
EXPERIMENTAL. Obtain adjusted and unadjusted PSM hazards for given endpoint and time
}
