% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmeans.R
\name{prmd_pd_stm_cf}
\alias{prmd_pd_stm_cf}
\title{Safely calculate restricted mean duration in progressed disease state for clock forward state transition model}
\usage{
prmd_pd_stm_cf(...)
}
\arguments{
\item{...}{Pass-through to \link{rmd_pd_stm_cf}}
}
\value{
Numeric value in same time unit as patient-level data (weeks).
}
\description{
Calculates the mean duration in the progressed disease state for the clock forward state transition model. Requires a carefully formatted list of fitted survival regressions for necessary endpoints, and the time duration to calculate over. Wrapper with 'possibly' of \link{rmd_pd_stm_cf}. This function is called by \link{calc_allrmds}.
}
