% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm3mkv-package.R
\docType{package}
\name{psm3mkv-package}
\alias{psm3mkv}
\alias{psm3mkv-package}
\title{psm3mkv: Fit and Evaluate Three-State Partitioned Survival Analysis and Markov Models to Progression-Free and Overall Survival Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fits and evaluates three-state partitioned survival analyses (PartSAs) and markov models (clock forward or clock reset) to Progression-Free Survival (PFS) and Overall Survival (OS) patient-level data. This data is typically collected in clinical trials in oncology. These model structures are typically considered in cost-effectiveness modeling in advanced/metastatic cancer indications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/psm3mkv/}
  \item \url{https://github.com/Merck/psm3mkv}
  \item Report bugs at \url{https://github.com/Merck/psm3mkv/issues}
}

}
\author{
\strong{Maintainer}: Dominic Muston \email{dominic.muston@merck.com} (\href{https://orcid.org/0000-0003-4876-7940}{ORCID})

Other contributors:
\itemize{
  \item Merck & Co., Inc. [copyright holder, funder]
}

}
\keyword{internal}
