% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm3mkv-package.R
\docType{package}
\name{psm3mkv-package}
\alias{psm3mkv}
\alias{psm3mkv-package}
\title{psm3mkv: Evaluate Partitioned Survival and State Transition Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fits and evaluates three-state partitioned survival analyses (PartSAs) and Markov models (clock forward or clock reset) to progression and overall survival data typically collected in oncology clinical trials. These model structures are typically considered in cost-effectiveness modeling in advanced/metastatic cancer indications. Muston (2024). "Informing structural assumptions for three state oncology cost-effectiveness models through model efficiency and fit". Applied Health Economics and Health Policy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/psm3mkv/}
  \item \url{https://github.com/Merck/psm3mkv}
  \item Report bugs at \url{https://github.com/Merck/psm3mkv/issues}
}

}
\author{
\strong{Maintainer}: Dominic Muston \email{dominic.muston@merck.com} (\href{https://orcid.org/0000-0003-4876-7940}{ORCID})

Other contributors:
\itemize{
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder, funder]
}

}
\keyword{internal}
